add expense seeds and change expense amount to float

+++ db/migrate/20170210062705_create_expenses.rb
@@ -2,7 +2,7 @@ class CreateExpenses < ActiveRecord::Migration
   def change
     create_table :expenses do |t|
       t.integer :user_id, null: false
-      t.integer :amount, null: false
+      t.float :amount, null: false
       t.string :description, null: false
       t.timestamps null: false
     end
